% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch sdmverse App
This function launch the sdmverse application.}
\usage{
launch_app(...)
}
\arguments{
\item{...}{additional parameters from the \code{\link[shiny:runApp]{runApp}} function in the \emph{shiny} package.}
}
\value{
This function does not return.
}
\description{
Launch sdmverse App
This function launch the sdmverse application.
}
\examples{
\dontrun{
  ForestryAnalysisInR::launchRFA(launch.browser = TRUE)
}

}
