% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dendrogram_figure.r
\name{make_dendrogram_figure}
\alias{make_dendrogram_figure}
\title{Create a dendrogram figure showing similarity between SDM packages}
\usage{
make_dendrogram_figure(
  table,
  diff_method = "euclidean",
  dendro_method = "average",
  ...
)
}
\arguments{
\item{table}{Tibble: A \code{tibble} with columns as package capacities and rows as packages. This can be created using \code{\link{prep_table}}.}

\item{diff_method}{Character: Method for calculating distances between packages. See \code{\link[stats]{dist}}.}

\item{dendro_method}{Character: Method for calculating clusters to create the dendrogram. See \code{\link[stats]{hclust}}.}

\item{...}{Arguments to pass to \code{\link{fviz_dend}}. See \code{Examples} for suggestions.}
}
\value{
A \pkg{ggplot2} object (class \code{gg} and \code{ggplot}).
}
\description{
This function creates a dendrogram figure based on the similarity between SDM packages.
}
\examples{

library(ggplot2)

# fetch data
table <- prep_table()

# a rectangular dendrogram
rect <- make_dendrogram_figure(
   table = table,
   k = nrow(table),
   k_colors = 'jco',
   rect = TRUE, 
   rect_border = 'jco', 
   rect_fill = TRUE,
   type = 'rectangle',
   horiz = TRUE,
   main = 'Species distribution modeling software',
   sub = date()
)

rect

# a phylogenic plot
phylo <- make_dendrogram_figure(
   table = table,
   k = nrow(table),
   k_colors = 'jco',
   rect = TRUE, 
   rect_border = 'jco', 
   type = 'phylogenic',
   repel = TRUE,
   main = 'Species distribution modeling software',
   sub = date()
)

phylo

}
