% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendrogram.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot a dendrogram}
\usage{
plot_dendrogram(
  table,
  diff_method = "binary",
  dendro_method = "average",
  return_clust = FALSE,
  ...
)
}
\arguments{
\item{table}{Tibble: A \code{tibble} with columns as package capacities and
rows as packages. This can be created using \code{\link{prep_table}}.}

\item{diff_method}{Character: Method for calculating distances between
packages. See \code{\link[stats]{dist}}.}

\item{dendro_method}{Character: Method for calculating clusters to create the
dendrogram. See \code{\link[stats]{hclust}}.}

\item{return_clust}{bool. Should the cluster figure be added.}

\item{...}{Arguments to pass to \code{\link{fviz_dend}}. See \code{Examples}
for suggestions.}
}
\value{
A \pkg{ggplot2} object (class \code{gg} and \code{ggplot}).
}
\description{
This function plot a dendrogram figure based on the similarity between SDM
packages.
}
\examples{
\donttest{
library(dplyr)
d <- prep_table(where = "online") \%>\%
  filter(
    name != "rgbif",
    name != "ibis.iSDM",
    name != "dismo"
  ) \%>\%
  select(
    -mod_mechanistic,
    -mod_multispecies
  )
plot_dendrogram(
  d,
  k = 5,
  cex = 0.7,
  diff_method = "binary",
  horiz = TRUE,
  main = ""
)
}

}
