% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Plot sdmverse table}
\usage{
plot_table(type = "tile")
}
\arguments{
\item{type}{char. Type of ggplot to make.}
}
\value{
Plot package characteristics from the sdmverse table using ggplot.
}
\description{
Plot sdmverse table
}
\examples{
\dontrun{
plot_table()
}

}
