% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_table.R
\name{prep_table}
\alias{prep_table}
\title{Prepare sdmverse packages table
This function prepare sdmverse packages table.}
\usage{
prep_table()
}
\value{
sdmverse packages metadata.
}
\description{
Prepare sdmverse packages table
This function prepare sdmverse packages table.
}
\examples{
\donttest{
prep_table()
}

}
