% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_cran.R
\name{scrape_cran}
\alias{scrape_cran}
\title{Scrape CRAN
This function scrape CRAN to verify package metadata}
\usage{
scrape_cran()
}
\description{
Scrape CRAN
This function scrape CRAN to verify package metadata
}
\examples{
\dontrun{
scrape_cran()
}

}
