% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmverse.R
\docType{package}
\name{sdmverse}
\alias{sdmverse}
\alias{sdmverse-package}
\title{sdmverse: A Meta-Package For Harnessing The Growing Diversity Of Species
Distribution Modeling Packages}
\description{
\code{sdmverse} is an annotated catalog of R packages for modeling species
distributions and ecological niches. Metadata on each package is contributed
by the package author(s). Please see
\href{https://github.com/sylvainschmitt/sdmverse}{instructions} for how to
contribute or update your package.
}
\details{
Create an issue on
\href{https://github.com/sylvainschmitt/sdmverse/issues}{GitHub}.
}
\section{Functions}{
 \code{\link{check_cran_metadata}}: Check package metadata
  with the metadata on CRAN\cr \code{\link{launch_app}}: Launch the
  \code{sdmverse} Shiny app\cr \code{\link{list_packages}}: Fetch a list of
  packages from the online \code{sdmverse} repository\cr
  \code{\link{plot_dendrogram}}: Cluster diagrams of packages based on their
  functionality\cr \code{\link{plot_table}}: Table of functionality of
  packages\cr \code{\link{prep_table}}: Fetch table of package metadata from
  online repository\cr \code{\link{template_from_cran}}: Create a \code{yaml}
  template for a package from CRAN metadata\cr
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sylvainschmitt/sdmverse}
  \item Report bugs at \url{https://github.com/sylvainschmitt/sdmverse/issues}
}

}
\author{
\strong{Maintainer}: Sylvain Schmitt \email{sylvain.m.schmitt@gmail.com} (\href{https://orcid.org/0000-0001-7759-7106}{ORCID})

Authors:
\itemize{
  \item Jamie Kass \email{jamie.m.kass@gmail.com} (\href{https://orcid.org/0000-0002-9432-895X}{ORCID})
}

Other contributors:
\itemize{
  \item Adam Smith \email{adam.smith@mobot.org} (\href{https://orcid.org/0000-0002-6420-1659}{ORCID}) [contributor]
}

}
\keyword{internal}
