% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_from_cran.R
\name{template_from_cran}
\alias{template_from_cran}
\title{Template from CRAN
This function create a yaml template from CRAN metadata.}
\usage{
template_from_cran(name, path = file.path("inst", "extdata", "packages"))
}
\arguments{
\item{name}{str. Name of the package on CRAN.}

\item{path}{str. Path to save the yaml file.}
}
\value{
Write the corresponding file in inst/extdata/pacakges.
}
\description{
Template from CRAN
This function create a yaml template from CRAN metadata.
}
\examples{
\dontrun{
template_from_cran()
}

}
