% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_cran_metadata.R
\name{check_cran_metadata}
\alias{check_cran_metadata}
\title{Check CRAN metadata
This function check package metadata with CRAN metadata.}
\usage{
check_cran_metadata()
}
\value{
TRUE or an error if this is not matching
}
\description{
Check CRAN metadata
This function check package metadata with CRAN metadata.
}
\examples{
\dontrun{
check_cran_metadata()
}

}
