% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch sdmverse App
This function launch the sdmverse application.}
\usage{
launch_app(
  port = getOption("shiny.port"),
  host = getOption("shiny.host", "127.0.0.1"),
  working.directory = getwd()
)
}
\arguments{
\item{port}{char. The TCP port that the application should listen on (see
\code{\link[shiny]{runApp}} for more details).}

\item{host}{char. The IPv4 address that the application should listen on (see
\code{\link[shiny]{runApp}} for more details).}

\item{working.directory}{char. Directory in which the application will run.}
}
\value{
Open a window with a shiny app to use the SSDM package with an
  user-friendly interface.
}
\description{
Launch sdmverse App
This function launch the sdmverse application.
}
\examples{
\dontrun{
  launch_app()
}

}
