% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_packages.R
\name{list_packages}
\alias{list_packages}
\title{List sdmverse packages
This function list sdmverse packages from online yamls.}
\usage{
list_packages(where = "online")
}
\arguments{
\item{where}{char. Where to download the packages metadata,
either 'locally' with the package or 'online' in the GitHub repository.}
}
\value{
sdmverse packages metadata.
}
\description{
List sdmverse packages
This function list sdmverse packages from online yamls.
}
\examples{
\donttest{
list_packages()
}

}
