% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dendrogram.R
\name{plot_dendrogram}
\alias{plot_dendrogram}
\title{Plot a dendrogram}
\usage{
plot_dendrogram(
  table,
  diff_method = "binary",
  dendro_method = "average",
  return_clust = FALSE,
  ...
)
}
\arguments{
\item{table}{Tibble: A \code{tibble} with columns as package capacities and
rows as packages. This can be created using \code{\link{prep_table}}.}

\item{diff_method}{Character: Method for calculating distances between
packages. See \code{\link[stats]{dist}}.}

\item{dendro_method}{Character: Method for calculating clusters to create the
dendrogram. See \code{\link[stats]{hclust}}.}

\item{return_clust}{bool. Should the cluster figure be added.}

\item{...}{Arguments to pass to \code{\link{fviz_dend}}. See \code{Examples}
for suggestions.}
}
\value{
A \pkg{ggplot2} object (class \code{gg} and \code{ggplot}).
}
\description{
This function plot a dendrogram figure based on the similarity between SDM
packages.
}
\examples{
\donttest{
# fetch data
table <- prep_table()

# a rectangular dendrogram
rect <- plot_dendrogram(
  table = table,
  dist_method = "binary",
  k = nrow(table),
  k_colors = "jco",
  rect = TRUE,
  rect_border = "jco",
  rect_fill = TRUE,
  type = "rectangle",
  horiz = TRUE,
  main = "Species distribution modeling software",
  sub = date(),
  cex = 1.4
)

# a phylogenic plot
phylo <- plot_dendrogram(
  table = table,
  dist_method = "binary",
  k = nrow(table),
  k_colors = "jco",
  rect = TRUE,
  rect_border = "jco",
  type = "phylogenic",
  repel = TRUE,
  main = "Species distribution modeling software",
  sub = date(),
  cex = 1
)
}

\dontrun{
library(patchwork)
rect + phylo
}

}
