% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.R
\name{plot_table}
\alias{plot_table}
\title{Plot sdmverse table}
\usage{
plot_table(d, pkg_order = NULL, pkg_cols = NULL, remove_empty_cats = FALSE)
}
\arguments{
\item{d}{data frame. sdmverse table built with prep_table().}

\item{pkg_order}{char. Order of packages to plot.}

\item{pkg_cols}{char. Color for packages plot.}

\item{remove_empty_cats}{bool. Remove empty categories.}
}
\value{
Plot package characteristics from the sdmverse table using ggplot.
}
\description{
Plot sdmverse table
}
\examples{
\donttest{
library(dplyr)
d <- prep_table(where = "online") \%>\%
  filter(
    name != "rgbif",
    name != "ibis.iSDM",
    name != "dismo"
  ) \%>\%
  select(
    -mod_mechanistic,
    -mod_multispecies
  )
plot_table(d, remove_empty_cats = TRUE)
}

}
