% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_table.R
\name{prep_table}
\alias{prep_table}
\title{Prepare sdmverse packages table
This function prepare sdmverse packages table.}
\usage{
prep_table(where = "online")
}
\arguments{
\item{where}{char. Where to download the packages metadata,
either 'locally' with the package or 'online' in the GitHub repository.}
}
\value{
sdmverse packages metadata.
}
\description{
Prepare sdmverse packages table
This function prepare sdmverse packages table.
}
\examples{
\donttest{
library(dplyr)
d <- prep_table(where = "online") \%>\%
  filter(
    name != "rgbif",
    name != "ibis.iSDM",
    name != "dismo"
  ) \%>\%
  select(
    -mod_mechanistic,
    -mod_multispecies
  )
}

}
